# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountant Practice Backoffice API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  code: string;
  name: string;
  type: string;
  /** @nullable */
  addressLine1?: string | null;
  /** @nullable */
  addressLine2?: string | null;
  /** @nullable */
  town?: string | null;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  postcode?: string | null;
  /** @nullable */
  contactNumber?: string | null;
  /** @nullable */
  email?: string | null;
}