# lib/api-zod/src/generated/types/financialInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountant Practice Backoffice API
 * OpenAPI spec version: 0.1.0
 */

export interface FinancialInfo {
  id: number;
  clientId: number;
  /** @nullable */
  clientCode?: string | null;
  /** @nullable */
  clientType?: string | null;
  /** @nullable */
  turnover?: number | null;
  /** @nullable */
  profitBeforeTax?: number | null;
  /** @nullable */
  totalIncome?: number | null;
  /** @nullable */
  totalSelfEmploymentIncome?: number | null;
  /** @nullable */
  totalProfitFromSelfEmployments?: number | null;
  /** @nullable */
  avgEmployees?: number | null;
}