# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountant Practice Backoffice API
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  /** @nullable */
  clientId: number | null;
  clientName?: string;
  /** @nullable */
  activityType?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  taskName: string;
  taskStatus: string;
  /** @nullable */
  dueDate?: Date | null;
  isOverdue?: boolean;
}