# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountant Practice Backoffice API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskInput {
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  activityType?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  taskName: string;
  taskStatus: string;
  /** @nullable */
  dueDate?: string | null;
}