# lib/api-zod/src/generated/types/taxReturn.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountant Practice Backoffice API
 * OpenAPI spec version: 0.1.0
 */

export interface TaxReturn {
  id: number;
  /** @nullable */
  clientId: number | null;
  /** @nullable */
  clientCode?: string | null;
  clientName?: string;
  /** @nullable */
  clientType?: string | null;
  /** @nullable */
  taxReturnStatus?: string | null;
}